web3_py_simple_storage/requirements.txt
py-solc-x == 1.1.1
web3 == 5.28.0
python-dotenv == 0.19.2